Gizbo MoveTo Command (WIP)
Addressed various bugs and amended parameters, according to recent changes in documentation. This includes:

Simplifying the implementation of the tasks / behaviours associated with Gizbo's commands;
Restricting the 'Move To' telegraph to be within a specified radius of the player;
Adding a Smart Link Nav Proxy to Nick's office, allowing Gizbo to actually get in and out of the room.

For further details, please see the Programming: Gizbo Commands task on ClickUp.